refactored services and tests
