+++ src/main/java/pingis/services/entity/TaskInstanceService.java
@@ -1,11 +1,9 @@
-package pingis.services;
+package pingis.services.entity;
 
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
-import java.util.Random;
-import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,43 +11,14 @@ import pingis.entities.Challenge;
 import pingis.entities.CodeStatus;
 import pingis.entities.Task;
 import pingis.entities.TaskInstance;
-import pingis.entities.TaskType;
 import pingis.entities.User;
 import pingis.repositories.TaskInstanceRepository;
-import pingis.repositories.TaskRepository;
-import pingis.repositories.UserRepository;
 
 @Service
 public class TaskInstanceService {
 
   @Autowired
-  private TaskRepository taskRepository;
-  @Autowired
   private TaskInstanceRepository taskInstanceRepository;
-  @Autowired
-  private UserRepository userRepository;
-  @Autowired
-  private UserService userService;
-
-  public TaskInstance getCorrespondingTestTaskInstance(
-      TaskInstance implTaskInstance) {
-    return taskInstanceRepository
-        .findByTaskAndUser(
-            taskRepository
-                .findByIndexAndChallengeAndType(implTaskInstance.getTask().getIndex(),
-                    implTaskInstance.getTask().getChallenge(), TaskType.TEST),
-            userRepository.findById(0L).get());
-  }
-
-
-  public TaskInstance getCorrespondingImplTaskInstance(TaskInstance testTaskInstance) {
-    return taskInstanceRepository
-        .findByTaskAndUser(
-            taskRepository.findByIndexAndChallengeAndType(testTaskInstance.getTask()
-                    .getIndex(),
-                testTaskInstance.getTask().getChallenge(), TaskType.IMPLEMENTATION),
-            userRepository.findById(0L).get());
-  }
 
   public TaskInstance findOne(long taskInstanceId) {
     Optional<TaskInstance> opt = taskInstanceRepository.findById(taskInstanceId);
@@ -84,29 +53,6 @@ public class TaskInstanceService {
     return taskInstances;
   }
 
-  @Transactional
-  public boolean rate(int rating, long taskInstanceId) {
-    Optional<TaskInstance> taskInstanceOpt = taskInstanceRepository.findById(taskInstanceId);
-
-    if (!taskInstanceOpt.isPresent()) {
-      return false;
-    }
-
-    TaskInstance taskInstance = taskInstanceOpt.get();
-
-    if (taskInstance.isRated()
-        || taskInstance.getUser().getId() != userService.getCurrentUser().getId()) {
-      // Task has already been rated or the instance doesn't belong to the current user, ignore.
-      return false;
-    }
-
-    Task task = taskInstance.getTask();
-
-    taskInstance.setRating(rating);
-    task.addRating(rating);
-
-    return true;
-  }
 
   @Transactional
   public TaskInstance markAsDone(TaskInstance taskInstance) {
@@ -134,6 +80,9 @@ public class TaskInstanceService {
     return taskInstances;
   }
 
+  public TaskInstance findByTaskAndUser(Task task, User user) {
+    return taskInstanceRepository.findByTaskAndUser(task, user);
+  }
 
   public int getNumberOfDoneTaskInstancesInChallenge(Challenge challenge) {
     int count = 0;
@@ -151,14 +100,6 @@ public class TaskInstanceService {
     return taskInstanceRepository.findByTaskAndUser(task, user);
   }
 
-  public TaskInstance getRandomTaskInstance(Task task) {
-    List<TaskInstance> viableInstances = taskInstanceRepository.findByTask(task).stream()
-        .filter(i -> !i.getUser().equals(userService.getCurrentUser()))
-        .filter(i -> i.getStatus() == CodeStatus.DONE)
-        .collect(Collectors.toList());
-    return viableInstances.get(new Random().nextInt(viableInstances.size()));
-  }
-
   public TaskInstance getUnfinishedInstanceInChallenge(Challenge challenge, User player) {
     Optional<TaskInstance> unfinished = getAllByChallenge(challenge).stream()
         .filter(i -> i.getUser().equals(player))
@@ -170,27 +111,14 @@ public class TaskInstanceService {
     return null;
   }
 
-  public List<TaskInstance> getHistory() {
-    return userService.getCurrentUser().getTaskInstances().stream()
-            .sorted(new TaskInstanceTimestampComparator())
-            .collect(Collectors.toList());
-  }
-
-  public TaskInstance getLastUnfinishedInstance() {
-    Optional<TaskInstance> unfinished = getHistory().stream()
-            .filter(e -> e.getStatus() == CodeStatus.IN_PROGRESS)
-            .findFirst();
-
-    if (unfinished.isPresent()) {
-      return unfinished.get();
-    } else {
-      return null;
+  public TaskInstance getUnfinishedTaskInstance(Challenge challenge) {
+    List<TaskInstance> taskInstances = getAllByChallenge(challenge);
+    for (TaskInstance current : taskInstances) {
+      if (current.getStatus() == CodeStatus.IN_PROGRESS) {
+        return current;
+      }
     }
-  }
-
-  public boolean canPlayOrSkip(TaskInstance taskInstance) {
-    return taskInstance.getUser().equals(userService.getCurrentUser())
-        && taskInstance.getStatus() == CodeStatus.IN_PROGRESS;
+    return null;
   }
 
   public class TaskInstanceTimestampComparator implements Comparator<TaskInstance> {
@@ -202,5 +130,13 @@ public class TaskInstanceService {
         return 1;
       }
     }
+  }
+
+  public List<TaskInstance> findByTask(Task task) {
+    return taskInstanceRepository.findByTask(task);
+  }
+
+  public Optional<TaskInstance> findById(Long id) {
+    return taskInstanceRepository.findById(id);
   }
 }
