+++ src/test/java/pingis/services/ArcadeChallengeServiceTest.java
@@ -1,0 +1,5 @@
+package pingis.services;
+
+public class ArcadeChallengeServiceTest {
+
+}
