+++ src/test/java/pingis/services/sandbox/SandboxServiceTest.java
@@ -35,7 +35,7 @@ import pingis.entities.sandbox.Submission;
 import pingis.entities.sandbox.SubmissionStatus;
 import pingis.entities.sandbox.TestOutput;
 import pingis.repositories.sandbox.SubmissionRepository;
-import pingis.services.TaskInstanceService;
+import pingis.services.entity.TaskInstanceService;
 
 @RunWith(SpringRunner.class)
 @ContextConfiguration(classes = {SandboxService.class})
